refactor(v2): split tools/registry.ts by subject; fail the build on unused code
registry.ts was 704 lines holding four unrelated things: the tool description
table, result rendering, the audited execution pipeline, eleven tool handlers,
and the MCP resource registrations. Everything shared one 500-line closure, so
any handler could reach the registry, policy engine or audit store directly and
quietly sidestep the pipeline.

Split by subject, largest file now 267 lines:

  descriptions.ts   TOOL_DESCRIPTIONS + getToolHashes
  results.ts        ToolResult, textResult, commandOutput, syntheticSuccess
  pipeline.ts       createPipeline() — policy, approval, audit, runAudited
  session-tools.ts  list-connections, list-sessions, open/close-session,
                    read-session-output
  command-tools.ts  read-command, run-command, privileged-command,
                    signal-process
  file-tools.ts     sftp-upload, sftp-download
  resources.ts      registerResources
  registry.ts       wiring only (35 lines), re-exporting the previous entry
                    points so importers are unchanged

The pipeline is a factory because it closes over server/registry/policy/audit.
Handler groups receive the built pipeline plus only the deps they use — after
the split the compiler showed file-tools needs nothing but `server`, and
session-tools nothing but `server` and `registry`. A handler wanting to bypass
policy or audit would now have to be handed something it isn't given, which is
the property the old single closure could not express.

Enables noUnusedLocals/noUnusedParameters in tsconfig.test.json, so `npm run
typecheck` fails on dead code rather than letting it accumulate — the same gap
that hid seven type errors until the test typecheck was added. It immediately
found pre-existing dead code beyond the split: unused imports in audit/store,
index, tracer, engine and four test files, and a private `outputBuffer` field in
InteractiveSession that was written on every run and never read.

Pure refactor: no behaviour change, verified by the unchanged suite —
248/248 across three consecutive runs.
