+++ src/tools/descriptions.ts
@@ -1,0 +1,23 @@
+import { createHash } from 'crypto';
+
+export const TOOL_DESCRIPTIONS: Record<string, string> = {
+  'list-connections': 'List all configured SSH profiles and their connection status. Use this to discover available hosts before running commands.',
+  'list-sessions': 'List active sessions for a given SSH profile.',
+  'open-session': 'Open a named session on a remote host. Use type="interactive" for stateful shell (CWD/env persists between commands) or type="background" for long-running processes.',
+  'close-session': 'Close a named session, releasing its resources.',
+  'read-session-output': 'Read recent output from a background session (e.g., tail -f logs).',
+  'read-command': 'Execute a READ-ONLY command from an allowlist (ls, cat, grep, find, stat, df, etc.). This tool does NOT modify the system. Prefer this tool for all read operations.',
+  'run-command': 'Execute an arbitrary shell command on the remote server. May modify the system. Destructive commands require user approval.',
+  'privileged-command': 'Execute a command with sudo elevation. ALWAYS requires user approval. The sudo password is piped via stdin (never visible in process list).',
+  'sftp-upload': 'Upload a file to the remote server via SFTP (secure file transfer, not shell-based).',
+  'sftp-download': 'Download a file from the remote server via SFTP.',
+  'signal-process': 'Send a signal (INT, TERM, KILL) to a remote process by PID.',
+};
+
+export function getToolHashes(): Record<string, string> {
+  const hashes: Record<string, string> = {};
+  for (const [name, desc] of Object.entries(TOOL_DESCRIPTIONS)) {
+    hashes[name] = createHash('sha256').update(desc).digest('hex').slice(0, 16);
+  }
+  return hashes;
+}
