+++ test/unit/audit/store.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import { mkdtemp, rm, readFile, stat, writeFile } from 'fs/promises';
+import { mkdtemp, rm, readFile, stat } from 'fs/promises';
 import { tmpdir } from 'os';
 import { join } from 'path';
 import { AuditStore } from '../../../src/audit/store.js';
